requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv>=1.0.0
